hdl/sm_io/*: add generic pointer to a specific handler
SMIO modules often need some state information.
Because of this, we store that in a new a generic
field called smio_handler, inside the general sm_io
structure. Specific realizations of the SMIO module
need to set that pointer to the desired structure,
initilize it correctly (inside bootstrap init function)
and destroy it when donw (inside bootstrap shutdown
function)