hal/dev_io/dev_io_core.*: add new poll methods
There are 2 new methods for doing the polling on
dev_io sockets. They are named devio_poll2_all_sm ()
and devio_init_poller2_sm () to differentiate from
the old methods and it performs a balanced polling.
For now, both of them are going to coexist. In the
future the old methods might be removed.