hal/dev_mngr/dev_mngr.c: add generic log filename
Now, we can pass the log directory and the dev_mngr
puts together the dev_mngr and dev_io log files
based on the number of devices currently running.

For now, the design is very simple and subject to
errors, such as when a dev_io exits.

In this case, the variable __dev_nums would be
decremented, but we wouldn't know which of the
dev_ios instances exited. The next time a dev_io
is started, the last dev_io logfile and device name
"/dev/fpga%d" would be choose for it, causing
the logs to be corrupted as well as PCIe access.