hal/dev_mngr/*: add new methods for spawning DEVIOs instances
Also, a few wrappers and trivial fixes were made.
Now, on dev_mngr initialization we scan the /dev
directory for our devices (/dev/fpga%d) and fill
the dev_info structures. After this, we can monitor
and spawn all of the DEVIOs that are not initialized.

This is another step in fixing the github issue #1.