hal/**/afcv3/*: add new static register layout for AFCv3
Also, we have changed the way we specify the devices. Now,
we have only the offsets to a generic base address. In this way,
we can instantiate various SMIO with the same name service, apart
from an ID at the end, such as: BPM0:DEVIO:SWAP0, BPM0:DEVIO:SWAP1
and so on.