hal/sm_io/*/acq/sm_io_acq_exp.c: use structutre for block aquisition
Our exported function structure says we return the data_block
structure, but we were returning just the valid bytes. This is
fixed now, as we return the whole block (even with invalid bytes)
and an extra field indicating which of the bytes are indeed valid.

This will be propagated to the libclient in the next commit.

This fixes #44 github issue.