hal/dev_mngr/dev_mngr_dev_info.*: add info about devio
As we merged the dev_io main program back into a single
code, we need a way to pass additional arguments to it.
So, we declare an dev_io_type parameter, meaning what SMIOs are
spawned, and a smio_inst_id parameter, used for DEVIOs with a
single SMIO, but multiple instances of it.