hal/ll_io/ops/ll_io_pcie.c: temporary fix to PCIe timeout bug
It was added a a PCIe timeout detection, so we can try to
detect a invalid stream of bytes and reset the FPGA timeout counter.
On detecting the timeout, we just sleep some period and hope that
there is no other contention inside the FPGA firmware.

This is a first fix to github issue #73. The "correct" fix is
described there.