libclient/Makefile: remove FORCE target
This target was added, as the dev_io would compile the
libclient and use it internally. Then, if the user also
wanted to install the libclient, the Makefile would not
regenerate the libclient objects, causing a possible
unwanted reuse of objects from a previous compilation.

Now, if the user wants to generate a clean libclient,
it should use make mrproper && make <desired flags> && make install.