hal/hal_utils/dispatch_table.*: fix permissive "const" on disp_op
We were not respecting the const qualifier. Because of this,
free () would not work sometimes and cause segfault.

In order to fix this, we used a wrapper to wrap the const disp_op
and the return value "ret". Some functions using disp_op were
changed to use disp_op_handler instead, but no semantic changes
were made and the API remains untouched.