include/hw/pcie_regs.h: fix RW block macro
The RW block macro wasn't working for sizes not multiple
of sizeof (BAR4), which is usually 8. In this way, reading/
writing 32-bit data, we were, in fact, reading/writing
nothing.

Also, we only do 32-bit transactions, even in BAR4. So,
we changed sizeof (*barp) to sizeof (BAR_RW_TYPE) which
is always 32-bit.

This is a first step into solving #12 github issue.