src/*/libbpmclient/src/*: fix request message wrong field size
On sending the message, the second parameter is not used, but as
we use the same message structure for read/write messages, the
server will still check for strict consistency of arguments.
Thus, for functions that expect a double argument, enconding a
uint32_t will result in an error.

The fix is simple, however. We just need to use the passed parameter
size to the message.