libs/libbpmclient/*: fix wrong pointer arithmetic
For the generic bpm_func_exec function and some
compounded libbpmclient funcions, we were incrementing
a uint32_t pointer by 4, effectively adding 4*4 = 16 bytes
to the pointer and not 4 bytes as intended. Everything seemed
to work OK, because the error was made twice: one filling
the array and the other reading the array (in bpm_func_exec).