sm_io/*/acq/*: fix acquisition address wrap
The acquisition address can wrap in two places.

The first of them is when calculation the start
of the acquisition, which could be wrapped if
the acquisition module waited too long for the
trigger.

The second place is in adding the block offset
to the start address. This could wrap in the end
of the memory space.