src/dev_io/dev_io.c: don`t try to get board slot for RFFE DEVIO
RFFE DEVIO already has the correct slot number
from the regular DEVIO. As such, it does not need
to spawn a libclient instance to get ther board
number from DEVIO CFG.

This simplifies the creataion/destruction of DEVIO
CFG, as we don`t need to wait for anything to destroy
it. The regular DEVIO just kills it after getting the
board number.

This fixes #94 github issue.