dev_io/dev_io_core.c: fix registering socket before creating it
As an optimization, we registered the SMIO message
and SMIO config handlers before creating the actor.

For the first case this is OK, as as register the
handler to an existing socket creted by
zsys_create_pipe ().

For the second case the registered socket does not
yet exist, as it will be creted by zactor_new ().