dev_io/dev_io_core.c: fix double socket reader removal on *_cfg handle
Removing the socket handler on *_cfg handle was wrong
as the devio_destroy_smio () already calls
devio_destroy_actor (), which, in turn, removes
the socket from the reader.

Also, as we expect only a single message on the
handler, it`s safer and cleaner to not poll the socket
until there is no message.