src,include/*: standardize the way we spawn ebpm application
Now, instead of spawning a DEVIO BE application
and that application, in turn, spawn the DEVIO FE,
we are using a more modular approach in that each
DEVIO BE or DEVIO FE spawns a single application.

We will rely on the system integration (such as
systemd) to take care of our dependencies for us.