dev_io/dev_io_core.c: don't create SDB if device not supported
As of now, devices that do not support SDB will fail,
as it does not implement a timeout mechanism.

For instance, Ethernet devices will wait forever to
receive the SDB magic number, but this will never happen.
So, avoid creating SDB in the first place until we have
a proper timeout/failover mechanism.