sm_io/protocols/*: restrict ops pointer access to inner class
In this way, the only option to get the operations
table for a particular "protocol" is to instantiate
the appropriate class and use the *_get_ops ()
function.

With this, we can be safe that the user will always
pass the pointer ops that is inside the "protocol"
structure (along with its options).