revision: add to Gradle build
Move source files to follow the Gradle convention. Create a build.gradle
script that declares the library component, specifies its dependency to
other libraries and exports the library's public headers.

Use the org.ajoberstar.release-opinion plugin to infer the project
version and use the GitConfigPlugin to extract the Git user information
to add it to the build as compiler defines.