gradle: create NativeDistributionPlugin
Creates a distribution for every build type, platform, flavor and usage
combination. Allows creating a RPM for every distribution using the
Redline RPM library.

RPMs used for development include only headers and depend on the
respective RPM used in runtime. Dependency between sub-projects is
handled by the RPM task, and the rule is to include a dependency to the
other sub-project's RPM if it exists, otherwise include the necessary
files from the other sub-project.