sm_io/*/fmc250m_4ch/*: remove temporary sleep/wait to timing issues
These were inserted as a workaround for timing
issues, in that we needed the ACTIVE_CLK
SMIO to be initialized first, but we had little
control on the timing of SMIO initialization.

However, with the github issue #125 fixed,
we can rely on this on not depend on sleep/wait
calls.