sm_io/*/sm_ch_24aa64.c: fix wrong pointer data passing
Previously, we were passing the address of the pointer
to the SMPR write function. Because of the explicity
cast this was not caught by the compiler, but this was
wrong and not what we wanted.

It was done like this, because before we used to have
an intermediate buffer to write data to the 24AA64. Then,
it made sense to do something like &data. Now, we don't
need this, but the code conversion was poorly done, so we
fix it here.

This led to errors on writing data to 24AA64.