gradle: allow uninstalling RPMs
Create an uninstall task that simply calls rpm -e for the specified
package. In order to properly set the dependencies, model rules were
created in order to collect all packages from all projects so that the
reverse dependency tree can be obtained. This tree is then used to
determine the uninstallation order.