core/*/rw_param.h: fix final address calculation by MACROS
Previously we were just ORing the base
address with the register offset. This
worked ok, as far as we kept the base address
on the MSB. But, recently, we decided to use
this base address as an offset for arrays of
registers, like trigger_iface/trigger_mux modules
that use a set of 3 registers replicated by
24 channels.

Now, to fix this, we just sum the base address
with the register offset. This works on both
cases (base address on MSB and on LSB).

This fixes #147 github issue.