libs/halcsclient: increase default timeout to 5000ms
Even though the default should be something less,
like 1000ms, this works better in the case of
SMIO initialization, for now.

This msut be fixed in some better way, but using 5000ms
as the defualt timeout is reasonable and can b cahnged
is the user wants, by using the *_new_log_log ()
client functions.