libs/llio: add new PCIe reset operation
This can be used to reset the PCIe
endpoint prior to other operation.

For example, on a late-initialization
procedure in that all of the SMIOs
are ready to go and the PCIe endpoint
needs to be reset prior to configuring
its defaults.

So, the steps would be the following in
this case:

1) Initialize the SMIOs
2) Reset endpoint
3) Configure the defaults