core/dev_io/*: fix UNREGISTER_SMIO command
Now, we can send the UNREGISTER_SMIO
to unregister a single SMIO.

Previously, the message was not conforming to
the actor received parameters so it was ignored.
Now, we've added a different pictore to the
zsock_recv () (from s484 to s484s) and we can
parse message in all cases.

In cases the message is shorter (all of them
except fro UNREGISTER_SMIO command) the pointers
are just nullified and no harm is done.