core/sm_io/*: use a small timeout of 100ms before exiting SMIO config
We were using a larger timeout os 2s, as we
relied on this to send amessage to the parent
application telling we have done configuring
all of the registered SMIOs.

Now, we don't need this anymore as we have fixed
some SMIOs initialization issues.

This timeout is only used to wait fence the parent
application before all SMIOs have configured
themselves.