core/*:/smio_thsafe_zmq_server.h: fix wrong structure for zmq_server_data_block
As we couldn't use a flexible array member
alone in the structure we had a dummy data
at the first bytes of the structure.

However, on receiving the block, we didn't
consider these first bytes as another variable,
causing decoding errors.

Simply passing this vartiable into the data block
would cause a performance bottleneck as we would
need to copy the whole user buffer to a temporary
buffer just to append these dummy data first.

A simpler approach was to use the old-fashioned
data[1] idiom to a flexbible array member.