sm_io/*/acq: fix wrong data block partitioning when address wrap
When acquiring data from triggered acquisitions
we can have a situation in which part of the data is
contained at the end of the memory region and the rest
is located at the beggining of it (address wrap).

So, we need to take that into account when acquiring
a data block, as it can be split.

Previously, we were only considering the case where a
whole block is fully cotained at the end or beggining
of the memory region, not split.