core/dev_io/*: fix LLIO unsafe devio_print_* () method
Previously, this method was notusing the actor
API and just accessed the LLIO directly, which,
if called after SMIO initialization could causes
concurrent-access issues.

Now, the methos just sends a message to DEVIO
and it's then processed safely by the DEVIO thread.

This is the first set into fixing #161 github issue.