core/*/hw/wb_acq_core_regs.h: add new regs defs to acq. core
This is an update of the register map, as
lnls-dig/infra-cores@0233340049 and
lnls-dig/bpm-gw@838752de53 changed it.

This was done to allow the software to dynamically
identify the multishot RAM size of each acquisition
core, and be possible for the gatewaew to select
different sizes for each multishow RAM, as well.