libs/halcsclient: robustify the way the client receives messages
Even with the previous fix (github issue #107),
there are still some cases in which the receiving
messages confuses the client.

Previously, we just dropped the message that
wasn't for the specific receiving client,
but that didn't solve the issue, as there were
always going to be an extra message in front of
the correct one.

Now, we discard all messages up to the point
there is a match in the "service" string,
so we can receive a message that was intended
fir the corresponding client.