sm_io/acq: remove wrong requirement of minimum sample_size for pre_samples
This is an old requirement that is not needed
by the ACQ gateware module anymore.

Also, we were only "enforcing" this when the
number os samples was 0, not when it was 1, 2, 3,
...

This is related to #168 github issue, fixing
the first part of the issue (using only the
minimum necessary for pre_samples).