libs: fix pkg-config .pc files to be relocatable
Previously, the pkg-config .pc files had a hardcoded
prefix in them, which made them not suitable for
installing with different packagers (i.e., make install vs
cpack debian). In this case the installation prefixes
are respectively changed to /usr/local and /usr. However,
the .pc file was always created on "cmake .." step with
the prefix being set to CMAKE_INSTALL_PREFIX. Which, In turn,
was set to /usr/local independently of the method of
installation.